# swaggerV2.json
{
  "swagger": "2.0",
  "info": {
    "description": "RESTfull API baseada em node.js Com o objetivo da partilha de viagens e experiencias vividas, classificações e comentáriosentre utilizadores. Disponível em https://tppweb.herokuapp.com",
    "version": "2.0.0",
    "title": "Trips Together",
    "contact": {
      "email": "o.nosso.email@gmail.com"
    }
  },
...
  "produces": [
    "application/json"
  ],
  "host": "tppweb.herokuapp.com",
  "basePath": "/api/v2",
  "paths": {
    "/viagens": {
      "get": {
        "tags": [
          "Viagens"
        ],
        "summary": "Obter viagens",
        "description": "Obter todas as viagens ou filtrar entre todas as viagens",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "description": "Filtros predefinidos",
            "type": "string",
            "enum": [
              "publicas",
              "minhas",
              "todas"
            ]
          },
          {
            "name": "filtro",
            "in": "query",
            "description": "Conteúdo da pesquisa a ser efetuada nos campos [Pais], [Cidade] e [Descrição]",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Resposta com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Viagens"
        ],
        "summary": "Inserir uma viagem",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "idUtilizador",
            "in": "formData",
            "description": "ID Utilizador",
            "required": true,
            "type": "string"
          },
          {
            "name": "partilhado",
            "in": "formData",
            "description": "Viagem partilhada",
            "type": "boolean",
            "required": true,
            "default": false
          },
          {
            "name": "popular",
            "in": "formData",
            "description": "Viagem popular",
            "type": "boolean",
            "required": true,
            "default": false
          },
          {
            "name": "pais",
            "in": "formData",
            "description": "Pais de destino",
            "type": "string"
          },
          {
            "name": "cidade",
            "in": "formData",
            "description": "Cidade de destino",
            "type": "string"
          },
          {
            "name": "latitude",
            "in": "formData",
            "description": "Coordenadas de latitude",
            "type": "string"
          },
          {
            "name": "longitude",
            "in": "formData",
            "description": "Coordenadas de longitude",
            "type": "string"
          },
          {
            "name": "dataInicio",
            "in": "formData",
            "description": "Data de inicio da viagem",
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "dataFim",
            "in": "formData",
            "description": "Data de fim da viagem",
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "descricao",
            "in": "formData",
            "description": "Descrição da viagem",
            "type": "string"
          },
          {
            "name": "ficheiros",
            "in": "formData",
            "description": "imagem associada a viagem",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Viagem inserida com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
    "/viagens/{id}": {
      "get": {
        "tags": [
          "Viagens"
        ],
        "summary": "Obtem todos os dados de uma viagem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID Viagem",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Viagem obtida com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Viagens"
        ],
        "summary": "Atualiza uma viagem existente",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da viagem",
            "required": true,
            "type": "string"
          },
          {
            "name": "partilhado",
            "in": "formData",
            "description": "Viagem partilhada",
            "type": "boolean",
            "required": true,
            "default": false
          },
          {
            "name": "popular",
            "in": "formData",
            "description": "Viagem popular",
            "type": "boolean",
            "required": true,
            "default": false
          },
          {
            "name": "pais",
            "in": "formData",
            "description": "Pais de destino",
            "type": "string"
          },
          {
            "name": "cidade",
            "in": "formData",
            "description": "Cidade de destino",
            "type": "string"
          },
          {
            "name": "latitude",
            "in": "formData",
            "description": "Coordenadas de latitude",
            "type": "string"
          },
          {
            "name": "longitude",
            "in": "formData",
            "description": "Coordenadas de longitude",
            "type": "string"
          },
          {
            "name": "dataInicio",
            "in": "formData",
            "description": "Data de inicio da viagem",
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "dataFim",
            "in": "formData",
            "description": "Data de fim da viagem",
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "descricao",
            "in": "formData",
            "description": "Descrição da viagem",
            "type": "string"
          },
          {
            "name": "ficheiros",
            "in": "formData",
            "description": "imagem associada a viagem",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Viagem atualizada com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Viagens"
        ],
        "summary": "Remover uma viagem por id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da viagem",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Viagem removida com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
    "/viagens/utilizador/{idUtilizador}": {
      "get": {
        "tags": [
          "Viagens"
        ],
        "summary": "Filtra todas as viagens por utilizador",
        "parameters": [
          {
            "name": "idUtilizador",
            "in": "path",
            "description": "ID do utilizador",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Viages obtidas com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
    "/viagens/cidade/{cidade}": {
      "get": {
        "tags": [
          "Viagens"
        ],
        "summary": "Filtra todas as viagens por cidade",
        "parameters": [
          {
            "name": "cidade",
            "in": "path",
            "description": "Nome da cidade",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Viages obtidas com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Response": {
      "type": "object",
      "properties": {
        "ErroCorde": {
          "type": "string"
        },
        "Message": {
          "type": "string"
        },
        "Error": {
          "type": "object"
        },
        "Data": {
          "type": "object"
        }
      }
    }
  }
}